pyproject.toml
[tool.poetry]
name = "linak-controller"
version = "1.3.0-dev"
description = "Command line tool for controlling the Linak standing desks"
authors = ["Rhys Tyers"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.14"
aiohttp = "^3.8.4"
appdirs = "^1.4.4"
bleak = "^0.22.3"
PyYAML = "^6.0"

[tool.poetry.scripts]
linak-controller = 'linak_controller.main:init'
idasen-controller = 'linak_controller.main:init'

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"